requirements.txt
Flask==3.0.3
requests==2.32.3
gunicorn==22.0.0
openai==2.32.0
python-dotenv==1.0.1
